KeyboardInput and GamepadInput now use elapsed time for movement distance.
AIActor and Player state collections replaced with boolean properties.

git-svn-id: https://lulz.unfuddle.com/svn/lulz_repos@25 62f4d935-ece3-4f1a-ac47-83beac009040

+++ Trunk/MyFirstGame/Actor/Actor.cs
@@ -67,7 +67,7 @@ namespace MyFirstGame.GameObject
             
         }
 
-        public void MoveTo(int x, int y)
+        public void MoveTo(float x, float y)
         {
             this.Position = new Vector2(x, y);
         }
